wip -- grow collections when adding items
